requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# LLM - Ollama
ollama==0.4.7
langchain==0.3.14
langchain-ollama==0.3.0

# GitHub Integration
PyGithub==2.5.0

# ML / Data Processing
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1

# HTTP Client
httpx==0.28.1

# Templating
jinja2==3.1.5

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
httpx==0.28.1

# Dev Tools
ruff==0.8.6
